+++ packages/astro/src/runtime/server/jsx.ts
@@ -12,7 +12,6 @@ import {
 } from './index.js';
 import { HTMLParts } from './render/common.js';
 import type { ComponentIterable } from './render/component';
-import { createScopedResult, ScopeFlags } from './render/scope.js';
 
 const ClientOnlyPlaceholder = 'astro-client-only';
 
@@ -95,8 +94,7 @@ Did you forget to import the component or is it possible there is a typo?`);
 						props[key] = value;
 					}
 				}
-				const scoped = createScopedResult(result, ScopeFlags.JSX);
-				const html = markHTMLString(await renderToString(scoped, vnode.type as any, props, slots));
+				const html = markHTMLString(await renderToString(result, vnode.type as any, props, slots));
 				return html;
 			}
 			case !vnode.type && (vnode.type as any) !== 0: